+++ src/main/java/org/monthly/payslip/tax/TaxableIncome.java
@@ -1,16 +1,22 @@
 package org.monthly.payslip.tax;
 
+/**
+ * Domain objects represents income tax ranges.
+ */
 public class TaxableIncome {
 
-    private final int from;
-    private final int to;
+    private final long from;
+    private final long to;
 
-    public TaxableIncome(int from, int to) {
+    public TaxableIncome(long from, long to) {
         this.from = from;
         this.to = to;
     }
 
-    public boolean isWithinRange(int income) {
+    /**
+     * Returns true if provided income falls between from and to values, false otherwise.
+     */
+    public boolean isWithinRange(long income) {
         return from <= income && income <= to;
     }
 }
